initial PR feedback
